subtask(F-2-6-split-install-bdd-step-modules): Split monolithic install BDD step implementation into focused modules
